Update Petri net as soon as the graph is changed
The previous method of converting the graph to a Petri net only when saving
led to a lot of problems, specifically out-of-sync situations and the inability
to simulate or even validate elements. Now, any change to the graph immediately
changes the Petri net and replaces the graph with a new one created from the
Petri net, ensuring that the GUI remains in sync with the internal structure.

This implementation is rather hacky, and the Petri net to graph functionality is
still incomplete, but it's a start.